scripts/requirements.txt
openpyxl>=3.1
requests>=2.31
Pillow>=10.0
google-generativeai>=0.8
